[#346] Move generate config into a util function
