+++ backend/core/config.py
@@ -1,9 +1,6 @@
-import os
-from jsmin import jsmin
 from fastapi import FastAPI, Request, Response
 from middleware import decode_token
 from fastapi.responses import FileResponse
-from db.connection import SessionLocal
 
 from routes.user import user_route
 from routes.tag import tag_route
@@ -18,10 +15,7 @@ from routes.cpi import cpi_route
 from routes.visualization import visualization_route
 from routes.reference_data import reference_data_routes
 
-from models.business_unit import BusinessUnit
-from models.commodity_category import CommodityCategory
-from models.currency import Currency
-from models.country import Country
+from utils.generate_config import generate_config_file, JS_FILE
 
 
 app = FastAPI(
@@ -40,49 +34,7 @@ app = FastAPI(
     },
 )
 
-JS_FILE = "./config.min.js"
-
-
-def generate_config_file() -> None:
-    session = SessionLocal()
-    env_js = "var __ENV__={"
-    env_js += 'client_id:"{}"'.format(os.environ["CLIENT_ID"])
-    env_js += ', client_secret:"{}"'.format(os.environ["CLIENT_SECRET"])
-    env_js += "};"
-    min_js = jsmin("".join([env_js, ""]))
-    business_units = session.query(BusinessUnit).all() or []
-    if business_units:
-        business_units = [bu.serialize for bu in business_units]
-    commodity_categories = session.query(CommodityCategory).all() or []
-    if commodity_categories:
-        commodity_categories = [
-            cc.serialize_with_commodities for cc in commodity_categories
-        ]
-    currencies = (
-        session.query(Currency.abbreviation, Currency.country).distinct() or []
-    )
-    if currencies:
-        currencies = [
-            {"value": c[0], "label": c[0], "country": c[1]} for c in currencies
-        ]
-    countries = (
-        session.query(Country).filter(Country.parent.is_(None)).all()
-        or []  # noqa
-    )
-    if countries:
-        countries = [c.to_dropdown for c in countries]
-    min_js += "var master={};".format(
-        str(
-            {
-                "business_units": business_units,
-                "commodity_categories": commodity_categories,
-                "currencies": currencies,
-                "countries": countries,
-            }
-        )
-    )
-    with open(JS_FILE, "w") as jsfile:
-        jsfile.write(min_js)
+generate_config_file()
 
 
 # Routes register
@@ -118,7 +70,6 @@ def health_check():
     description="static javascript config",
 )
 async def main(res: Response):
-    generate_config_file()
     res.headers["Content-Type"] = "application/x-javascript; charset=utf-8"
     return JS_FILE
