+++ backend/generate_config.sh
@@ -1,0 +1,3 @@
+#!/usr/bin/env bash
+
+python -m utils.generate_config
